Initial (admittedly large) commit with honesty
As humans, we grow with honesty. This commit has been long overdue, emphasis on `long`. My commit history should have started with the initialization of this project and other projects too for that matter. I severely underestimated the value and the normality of having a documented public history of working in a repo. I don’t expect further commits to this project to undo the neglect of consistently pushing changes bit by bit...but still, better late than never. If you’re reading this, thank you. I appreciate ya’ll.